+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Make internal podcast handle unique which is now combination of {podcastId}-{podcastSlug} instead of {podcastSlug}
 - Each podcast has its heading in users permission page
 - Admin users can grant more specific permissions to users
+- Episode title, duration, number and pubDate are not overwritten when fetching metadata if they are not empty
 - Fixed route for episodes
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 - Fixed permission issues where non-admin users were not able to access some plugin actions